% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_db.R
\name{count_levels}
\alias{count_levels}
\title{Contar niveles de anotación}
\usage{
count_levels(keys, keytype = "R_Name")
}
\arguments{
\item{keys}{Lista de identificadores de la columna indicada en \code{keytype}.}

\item{keytype}{Cadena con el nombre de la columna que se desea usar como llave ("R_Name" por defecto).}
}
\description{
A partir de una lista de identificadores, devuelve un \emph{data frame} con el conteo de niveles de anotación
RefMet indicados en "Metabolitos.xlsx" para las filas correspondientes.
}
