% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{create_metabolite_db}
\alias{create_metabolite_db}
\title{Crear las bases de datos SQLite}
\usage{
create_metabolite_db()
}
\description{
A partir del archivo "Metabolitos.xlsx", crea una base de datos
SQLite ("metabolites.sqlite") con dos tablas, "metabolitos" (que es equivalente
a la tabla de "Metabolitos.xlsx") y "metabolitosExp" (una versión
expandida de "metabolitos" en la que en la que las observaciones con valores
múltiples han sido desglosadas en varias filas, de tal forma que cada
combinación única de identificadores se encuentra en una fila separada).
}
