% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_db.R
\name{filter_by}
\alias{filter_by}
\title{Filtrar por columna}
\usage{
filter_by(
  keys,
  keytype = "R_Name",
  columnas = "*",
  filtro,
  valor,
  igual = TRUE
)
}
\arguments{
\item{keys}{Lista de identificadores de la columna indicada en \code{keytype}.}

\item{keytype}{Cadena con el nombre de la columna que se desea usar como llave ("R_Name" por defecto).}

\item{columnas}{Lista de columnas que se desean extraer (por defecto, "*", es decir, todas).}

\item{filtro}{Cadena con el nombre de la variable por la que se quiere filtrar.}

\item{valor}{Valor de la variable \code{filtro} por el que se quiere filtrar.}

\item{igual}{Booleano que indica si \code{valor} debe excluirse (\code{FALSE}) o si solo deben incluirse las filas que contenga \code{valor} (TRUE).}
}
\description{
A partir de una lista de identificadores, permite filtrar las correspondientes filas de la base de datos
en función de la presencia o ausencia de un valor determinado de una columna determinada.
}
