% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_db.R
\name{get_id}
\alias{get_id}
\title{Obtener los valores separados de un identificador}
\usage{
get_id(keys, keytype = "R_Name", columnas = keytype)
}
\arguments{
\item{keys}{Lista de identificadores de la columna indicada en \code{keytype}.}

\item{keytype}{String con el nombre de la columna que se desea usar como llave ("R_Name" por defecto).}

\item{columnas}{Lista de columnas que se desean extraer (por defecto, la indicada en \code{keytype}).}
}
\description{
A partir de una lista de identificadores, permite obtener una tabla con las filas de "metabolitosExp"
correspondientes. Además, se pueden elegir las columnas que se desean obtener.

De esta forma, se pueden extraer todos los identificadores de interés de los metabolitos indicados
separados individualmente, incluso cuando alguno de ellos tenga varios identificadores del mismo tipo.
}
\seealso{
\code{\link[=id_table]{id_table()}}
}
