% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_db.R
\name{id_table}
\alias{id_table}
\title{Obtener identificadores}
\usage{
id_table(
  keys,
  keytype = "R_Name",
  columnas = c("Pr_Key", "Order_Num", "R_Name", "metabolite", "HMDB_ID", "KEGG_ID",
    "InChI_Key", "FOBI_ID")
)
}
\arguments{
\item{keys}{Lista de identificadores de la columna indicada en \code{keytype}.}

\item{keytype}{Cadena con el nombre de la columna que se desea usar como llave (por defecto, "R_Name").}

\item{columnas}{Lista de columnas que se desean extraer (por defecto,
\code{c("Pr_Key", "Order_Num", "R_Name", "metabolite", "HMDB_ID", "KEGG_ID", "InChI_Key", "FOBI_ID")})}
}
\description{
A partir de una lista de identificadores, permite obtener una tabla con las filas de "metabolitos"
correspondientes. Además, se pueden elegir las columnas que se desean obtener.
}
\seealso{
\code{\link[=get_id]{get_id()}}
}
