% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_db.R
\name{keytypes}
\alias{keytypes}
\title{Obtener los keytypes}
\usage{
keytypes()
}
\description{
Muestra los nombres de las llaves que aceptan los parámetros "keytype", "columnas" y "filtro"
de las funciones de este paquete.
}
