% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboliteDbi-package.R
\docType{package}
\name{MetaboliteDbi}
\alias{MetaboliteDbi-package}
\alias{MetaboliteDbi}
\title{MetaboliteDbi: Exploración de la Base de Datos de Metabolitos}
\description{
Permite resolver las ambigüedades presentes en la tabla de metabolitos "Metabolitos.xlsx" y analizarla
mediante consultas SQL.
}
\details{
Este paquete permite transformar la tabla de metabolitos "Metabolitos.xlsx" en una base de datos SQLite
con dos tablas: "metabolitos" (equivalente a "Metabolitos.xlsx" más una llave primaria) y "metabolitosExp",
una versión expandida de "metabolitos" en la que los identificadores múltiples han sido desglosados en filas
individuales.
Mediante consultas SQL a ambas tablas relacionadas, se pueden resolver las ambigüedades presentes en la tabla
de metabolitos. Para facilitar estas consultas, se incluyen funciones que realizan de forma automatizada algunas
de las consultas más comúnes: búsqueda de metabolitos, filtrados y el contaje de niveles de anotación (RefMet).
}
\author{
\strong{Maintainer}: José Manuel Guerrero Jiménez \email{jguerrerojimenez@uoc.edu}

}
